# openapi_sveta.yaml
openapi: 3.0.3
info:
  title: API spb4you 3.0
  version: 1.0.11
  description: Все необходимые API для сайта Питер_Для_Тебя
...
paths:
...
  /users/info/{userId}:
    get:
      tags:
        - UsersController
      summary: Получить информацию об определенном пользователе
      description: Возвращает данные пользователя по id
      parameters:
        - name: userId
          in: path
          required: true
          description: Уникальный идентификатор пользователя
          schema:
            type: integer
      responses:
        '200':
          description: Успешное получение информации о пользователе
        '404':
          description: Пользователь не найден
  /users/registry:
    get:
      tags:
        - UsersController
      summary: Получение формы регистрации
      description: Возвращает страницу с формой регистрации
      responses:
        '200':
          description: Страница с формой регистрации
        '500':
          description: Ошибка сервера

    post:
      tags:
        - UsersController
      summary: Регистрация нового пользователя
      description: Позволяет новому пользователю зарегистрироваться в системе.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: "Света"
                email:
                  type: string
                  description: Электронная почта пользователя
                  nullable: true
                  example: "sveta@gmail.com"
                password:
                  type: string
                  writeOnly: true
                  example: "666#999"
              required:  # Обязательные поля
                - username
                - email
                - password
      responses:
        '201':
          description: Успешная регистрация
        '400':
          description: Ошибка валидации данных
  /users/login:
    get:
      tags:
        - UsersController
      summary: Получение формы авторизации
      description: Возвращает страницу с формой для входа в аккаунт
      responses:
        '200':
          description: Страница с формой авторизации
        '500':
          description: Ошибка сервера

    post:
      tags:
        - UsersController
      summary: Авторизация пользователя
      description: Авторизует пользователя и перенаправляет на соответствующую главную страницу.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Света"
                password:
                  type: string
                  example: "666#999"
              required:  # Обязательные поля
                - username
                - password
      responses:
        '200':
          description: Успешная авторизация
        '401':
          description: Неверные учетные данные
  /users/index/{userId}:
    get:
      tags:
        - UsersController
      summary: Получить главную страницу пользователя
      description: Возвращает главную страницу пользователя по id
      parameters:
        - name: userId
          in: path
          required: true
          description: Уникальный идентификатор пользователя
          schema:
            type: integer
      responses:
        '200':
          description: Успешное получение главной страницы пользователя
  /users/cabinet/{userId}:
    get:
      tags:
        - UsersController
      summary: Получить страницу личного кабинета пользователя
      description: Возвращает страницу личного кабинета пользователя по id
      parameters:
        - name: userId
          in: path
          required: true
          description: Уникальный идентификатор пользователя
          schema:
            type: integer
      responses:
        '200':
          description: Успешное получение страницы личного кабинета пользователя
  /users/{userId}/like-locations/{locationId}:
    post:
      tags:
        - UsersController
      summary: Добавить лайк на локацию пользователем
      description: Добавляет лайк к указанной локации для данного пользователя по их id
      parameters:
        - name: userId
          in: path
          required: true
          description:  Уникальный идентификатор пользователя
          schema:
            type: integer
        - name: locationId
          in: path
          required: true
          description:  Уникальный идентификатор локации
          schema:
            type: integer
      responses:
        '200':
          description: Локация успешно добавлена в список избранных
          content:
            application/json:
              schema:
                type: string
                example: "Location liked."
        '400':
          description: Локация уже лайкнута пользователем или не найдена
          content:
            application/json:
              schema:
                type: string
                example: "Location already liked."
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                type: string
                example: "User not found."
        '500':
          description: Ошибка сервера
          content:
            application/json:
              schema:
                type: string
                example: "Internal server error."
  /users/{userId}/unlike-locations/{locationId}:
    post:
      tags: 
        - UsersController
      summary: Удалить лайк с локации пользователем
      description: Удаляет лайк с указанной локации для данного пользователя по их ID.
      parameters:
        - name: userId
          in: path
          required: true
          description: Уникальный идентификатор пользователя
          schema:
            type: integer
        - name: locationId
          in: path
          required: true
          description:  Уникальный идентификатор локации
          schema:
            type: integer
      responses:
        '200':
          description: Локация успешно удалена из списка избранных
          content:
            application/json:
              schema:
                type: string
                example: "Location unliked."
        '400':
          description: Локация не была лайкнута пользователем или не найдена
          content:
            application/json:
              schema:
                type: string
                example: "Location not liked yet."
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                type: string
                example: "User not found."
  /users/{userId}/like-locations:
    get:
      tags:
        - UsersController
      summary: Получить страницу с избранными локациями пользователя
      description: Возвращает страницу с локациями, которые пользователь добавил в избранное по его id
      parameters:
        - name: userId
          in: path
          required: true
          description: Уникальный идентификатор пользователя
          schema:
            type: integer
      responses:
        '200':
          description: Страница с избранными локациями пользователя
        '302':
          description: Страница ошибки, если пользователь не найден
        '500':
          description: Ошибка сервера
  /users/{userId}/likes:
    get:
      tags:
        - UsersController
      summary: Получить страницу "избранное" пользователя
      description: Возвращает страницу "избранное" по id пользователя
      parameters:
        - name: userId
          in: path
          required: true
          description: Уникальный идентификатор пользователя
          schema:
            type: integer
      responses:
        '200':
          description: Страница "избранное" пользователя
        '302':
          description: Страница ошибки, если пользователь не найден
        '500':
          description: Ошибка сервера
...