# src/main/java/com/example/spb4you/controllers/UsersController.java
package com.example.spb4you.controllers;

import com.example.spb4you.models.Location;
import com.example.spb4you.services.LocationService;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import com.example.spb4you.models.User;
import com.example.spb4you.services.UserService;
import jakarta.servlet.http.HttpSession;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Controller;
import org.springframework.ui.Model;
import org.springframework.web.bind.annotation.*;

import java.time.LocalDateTime;
import java.util.ArrayList;
import java.util.List;
import java.util.stream.Collectors;

@Controller
@RequestMapping("/users")
public class UsersController {

    @Autowired
    private UserService userService;

    @Autowired
    private LocationService locationService;

    @GetMapping("/info/list")
    public ResponseEntity<List<User>> listUsers() {
        List<User> users = userService.findAll();
        return ResponseEntity.ok(users); // Возвращаем список всех пользователей со всеми полями для каждого с кодом 200
    }

    @GetMapping("/info/{userId}")
    public ResponseEntity<User> getUserDetails(@PathVariable("userId") Integer userId) {
        User user = userService.findById(userId).orElse(null);
        assert user != null;
        return ResponseEntity.ok(user); // Возвращаем набор полей пользователя по соответствующему ID с кодом 200
    }

    @GetMapping("/registry")
    public String showRegistrationForm(Model model) {
        model.addAttribute("user", new User()); // Создание нового объекта User
        return "registration";
    }

    @PostMapping("/registry")
    public String registerUser(@ModelAttribute User user, Model model) {
        // Проверка на существующего пользователя
        if (userService.findByEmail(user.getEmail()).isPresent()) {
            model.addAttribute("errorMessage", "Ошибка: Пользователь с таким email уже существует.");
            return "registration"; // Возвращение на страницу регистрации с сообщением об ошибке
        }

        // Хэширование пароля (для безопасности)

        // Сохранение нового пользователя в БД
        user.setRegistrationDate(LocalDateTime.now());
        user.setLikedLocations(", ");
        user.setLikedRoutes(", ");
        userService.save(user);

        model.addAttribute("firstName", user.getFirstName()); // Сохраняем имя пользователя в модели
        // Перенаправление на личную страницу пользователя
        return "redirect:/users/index/" + user.getId(); // Используем ID для формирования URL
    }

    @GetMapping("/login")
    public String showLoginForm(Model model) {
        model.addAttribute("user", new User()); // Создаем новый объект User в модели
        return "auth"; // Возвращаем страницу входа
    }

    @PostMapping("/login")
    public String loginUser(@ModelAttribute User user, HttpSession session, Model model) {
        // Проверка на существующего пользователя
        User existingUser = userService.findByFirstName(user.getFirstName()).orElse(null);
// ВАЖНО - добавить обработку при совпадении имён
        if (existingUser == null || !existingUser.getPassword().equals(user.getPassword())) {
            model.addAttribute("errorMessage", "Ошибка: Неверное имя пользователя или пароль");
            return "auth"; // Возвращаем на страницу логина с сообщением об ошибке
        }

        // Сохраняем id пользователя в сессии
        session.setAttribute("userId", existingUser.getId());

        // Перенаправление на личную страницу пользователя
        return "redirect:/users/index/" + existingUser.getId(); // Используем ID для формирования URL
    }

    /*
    @GetMapping("/logout")
    public String logout() {
        return "redirect:/login"; // Перенаправление на страницу логина после выхода
    }
    */

    @GetMapping("/index/{userId}")
    public String userInfo(@PathVariable("userId") Integer userId, Model model) {
        User user = userService.findById(userId).orElse(null);
        if (user == null) {
            // Обработка случая, если пользователь не найден
            return "redirect:/error"; // Возвращаем ошибку
        }
        model.addAttribute("user", user);
        return "index"; // Главная страница зарегистрированного пользователя
    }

    @GetMapping("/cabinet/{userId}")
    public String userCabinet(@PathVariable("userId") Integer userId, Model model) {
        User user = userService.findById(userId).orElse(null);
        if (user == null) {
            // Обработка случая, если пользователь не найден
            return "redirect:/error"; // Возвращаем ошибку
        }
        model.addAttribute("user", user);
        return "cabinet"; // Личный кабинет зарегистрированного пользователя
    }

    @PostMapping("/{userId}/like-locations/{locationId}")
    public ResponseEntity<?> likeLocation(@PathVariable Integer userId, @PathVariable Integer locationId) {
        try {
            System.out.println("Attempting to like location ID: " + locationId + " for user ID: " + userId);

            // Получение пользователя
            User user = userService.findById(userId).orElse(null);
            if (user == null) {
                System.out.println("User not found!");
                return ResponseEntity.status(HttpStatus.NOT_FOUND).body("User not found.");
            }

            // ПолучениеLikedLocations
            List<Integer> likedLocations = user.getLikedLocationsList();
            if (likedLocations == null) {
                likedLocations = new ArrayList<>(); // Инициализация списка
            }

            System.out.println("Current liked locations: " + likedLocations);

            // Проверка наличия лайка
            if (!likedLocations.contains(locationId)) {
                likedLocations.add(locationId);
                user.setLikedLocationsList(likedLocations);
                userService.save(user); // Сохраняем пользователя

                // Попытка получить локацию
                Location location = locationService.findById(locationId).orElse(null);
                if (location != null) {
                    location.setLikes(location.getLikes() + 1);
                    locationService.save(location);
                    System.out.println("Location liked successfully.");
                    return ResponseEntity.ok("Location liked.");
                } else {
                    System.out.println("Location not found!");
                    return ResponseEntity.badRequest().body("Location not found.");
                }
            } else {
                System.out.println("Location already liked.");
                return ResponseEntity.badRequest().body("Location already liked.");
            }
        } catch (Exception e) {
            e.printStackTrace(); // Печать информации об ошибке
            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Internal server error.");
        }
    }

    @PostMapping("/{userId}/unlike-locations/{locationId}")
    public ResponseEntity<?> unlikeLocation(@PathVariable Integer userId, @PathVariable Integer locationId) {
        System.out.println("Attempting to unlike location ID: " + locationId + " for user ID: " + userId);

        User user = userService.findById(userId).orElse(null);
        if (user == null) {
            System.out.println("User not found!");
            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("User not found.");
        }

        // Получаем списокLikedLocations
        List<Integer> likedLocations = user.getLikedLocationsList();
        if (likedLocations == null) {
            likedLocations = new ArrayList<>(); // Возможно, список не был инициализирован
        }

        System.out.println("Current liked locations: " + likedLocations);

        // Логика удаления лайка с локации
        if (likedLocations.contains(locationId)) {
            likedLocations.remove(locationId);
            user.setLikedLocationsList(likedLocations);
            if (user.getLikedLocations() == null || user.getLikedLocations().trim().isEmpty()) {
                user.setLikedLocations(", ");
            }
            userService.save(user); // Сохраняем изменения пользователя

            // Уменьшаем параметр likes локации
            Location location = locationService.findById(locationId).orElse(null);
            if (location != null) {
                location.setLikes(location.getLikes() - 1);
                locationService.save(location); // Сохраняем изменения локации
                System.out.println("Location unliked successfully.");
                return ResponseEntity.ok("Location unliked.");
            } else {
                System.out.println("Location not found!");
                return ResponseEntity.badRequest().body("Location not found.");
            }
        }

        System.out.println("Location was not liked by the user.");
        return ResponseEntity.badRequest().body("Location not liked yet.");
    }

    @GetMapping("/{userId}/like-locations")
    public String userShowLocationsLikedPage(@PathVariable("userId") Integer userId, Model model) {
        User user = userService.findById(userId).orElse(null);
        if (user == null) {
            // Обработка случая, если пользователь не найден
            return "redirect:/error"; // Возвращаем ошибку
        }

        // Получаем список идентификаторов локаций, которые пользователь добавил в избранное
        List<Integer> likedLocationIds = user.getLikedLocationsList(); // Предполагаем, что это возвращает List<Integer>

        // Фильтруем локации, чтобы оставить только те, которые есть в избранном у пользователя
        List<Location> likedLocations = locationService.findAll()
                .stream()
                .filter(location -> likedLocationIds.contains(location.getId())) // Проверяем наличие id в списке
                .toList();

        model.addAttribute("likedLocations", likedLocations);
        model.addAttribute("userId", userId);
        return "likedLocationsPage"; // Страница с избранными локациями
    }

    @GetMapping("/{userId}/likes")
    public String userShowLikedPage(@PathVariable("userId") Integer userId, Model model) {
        User user = userService.findById(userId).orElse(null);
        if (user == null) {
            // Обработка случая, если пользователь не найден
            return "redirect:/error"; // Возвращаем ошибку
        }

        // Получаем список идентификаторов локаций, которые пользователь добавил в избранное
        List<Integer> likedLocationIds = user.getLikedLocationsList(); // Предполагаем, что это возвращает List<Integer>

        // Фильтруем локации, чтобы оставить только те, которые есть в избранном у пользователя
        List<Location> likedLocations = locationService.findAll()
                .stream()
                .filter(location -> likedLocationIds.contains(location.getId())) // Проверяем наличие id в списке
                .toList();

        model.addAttribute("likedLocations", likedLocations);
        return "likedPage"; // Страница с избранными локациями и маршрутами
    }
}